test: add folders and move test files
